Refactor to make separate CallAssignmentTargets component.
